pyproject.toml
[project]
name = "alpha-canvas"
version = "0.1.0"
description = ""
authors = [
    {name = "jaepil-choi.laptop",email = "chljeffreyz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest (>=8.4.2,<9.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "xarray (>=2025.10.1,<2026.0.0)",
    "duckdb (>=1.4.1,<2.0.0)",
    "pyarrow (>=21.0.0,<22.0.0)",
    "scipy (>=1.16.2,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "bottleneck (>=1.6.0,<2.0.0)",
    "curl-cffi (>=0.13.0,<0.14.0)",
    "ipykernel (>=7.1.0,<8.0.0)",
    "notebook (>=7.4.7,<8.0.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
]

[tool.poetry]
packages = [
    {include = "alpha_excel", from = "src"},
    {include = "alpha_database", from = "src"},
    {include = "alpha_lab", from = "src"},
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "src"
]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "-m", "not smoke"  # Skip smoke tests by default
]
markers = [
  "unit: Independent unit tests",
  "integration: Integration tests that test the interaction between components",
]